--- src/main/java/roomescape/dto/ReservationResponse.java
@@ -1,47 +1,0 @@
-package roomescape.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import roomescape.domain.Reservation;
-
-public class ReservationResponse {
-
-    private final long id;
-    private final String name;
-    private final LocalDate date;
-    @JsonFormat(pattern = "HH:mm")
-    private final LocalTime time;
-
-    private ReservationResponse(long id, String name, LocalDate date, LocalTime time) {
-        this.id = id;
-        this.name = name;
-        this.date = date;
-        this.time = time;
-    }
-
-    public static ReservationResponse from(Reservation reservation) {
-        return new ReservationResponse(
-            reservation.getId(),
-            reservation.getName(),
-            reservation.getDate(),
-            reservation.getTime()
-        );
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public LocalTime getTime() {
-        return time;
-    }
-}
